Some changes where an actual target was used.
Made sure that the interrupt and timer had basic functionality.

Used buttons and serial output to test though. Next steps would be
to confirm the state of PWMs and the ability to drive a servo.

My motion sensor SHOULD be indiscriminate from a button though...
so I'm not concerned about testing that.